requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
mac-vendor-lookup==0.1.12
multidict==6.0.4
Pillow==9.4.0
playsound==1.3.0
scapy==2.5.0
termcolor==2.2.0
yarl==1.8.2
# vext==0.7.6; sys_platform == 'linux'
# vext.gi==0.7.4; sys_platform == 'linux'